# Doc/openapi.yaml
openapi: 3.0.3
info:
  title: MarketCatalyst (FinApp26) Market Data API
  version: 1.0.0
  description: |
    Canonical read API for every screen in MarketCatalyst. This documents the **data
    contract** the frontend consumes — today implemented as direct Firestore
    reads (via `useCollection()` in `app/iq/hooks/useCollection.ts`) rather than
    a deployed REST gateway, so every path below maps 1:1 to a Firestore
    collection written by a NestJS sync job in `backend/src/sync/`. The shapes
    are copied verbatim from the actual job code (not from the earlier planning
    doc `Doc/05_API_Vendor_Tracker.md`, which has drifted from what is really
    implemented — e.g. `analyst_actions` was planned as a per-event
    upgrade/downgrade feed but is actually a consensus-vote snapshot today).

    **Status legend** (see `x-status` on every path):
    - `live` — job exists, writes real data, screen(s) already read it.
    - `planned` — no backend job exists yet; schema is a proposed shape based
      on the named vendor's typical payload. Nothing reads/writes this today.

    **Source legend** (see `x-primary-source` / `x-alternate-source` on every
    path): the vendor actually wired in `backend/src/vendors/`, plus a faster
    or more complete alternate and why it's faster — see each path's
    description for specifics.

    **Swapping a source never changes this document — and fallback is
    automatic and never silent.** `Company` and `MarketMover` (the two
...

...
  contact:
...
servers:
...
paths:
...
  /plans/seed:
    post:
      tags: [Plans]
      summary: Re-seed plans from the code registry (merge-based)
      description: |
        Merge-based: operator edits made in Firestore survive. Commercial fields
        (price/currency/cycle) are written only on FIRST create, so a re-seed
        never silently reverts a price change.
      x-status: live
      x-auth: AdminGuard
      operationId: seedPlans
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  seeded: { type: integer }
                  keysBackfilled: { type: integer }
        "401": { description: Not an admin }
  /users/{uid}/entitlements:
    get:
      tags: [Plans]
      summary: Effective entitlements for one user (release flag AND plan grant)
      description: |
        Returns BOTH gating axes separately plus their AND. `released` = the FF_*
        release flag ("is it built?"); `entitled` = the plan grant ("may this tier
        use it?"); `enabled` = both. `reason` distinguishes `not-released` from
        `not-in-plan` so the UI can say "coming soon" vs "upgrade to unlock" —
        collapsing them would make an unbuilt feature look like a paywall.

        Subscription expiry is COMPUTED here, never trusted from the stored
        status: nothing rewrites a user document at the moment a subscription
        lapses, so a stale `ACTIVE` would otherwise grant paid access forever.
      x-status: live
      x-auth: AdminGuard
      operationId: getUserEntitlements
      parameters:
        - name: uid
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription: { $ref: "#/components/schemas/EffectiveSubscription" }
                  entitlements:
                    type: array
                    items: { $ref: "#/components/schemas/ResolvedEntitlement" }
                  enabled:
                    type: object
                    additionalProperties: { type: boolean }
...
components:
...
  schemas:
...
    EffectiveSubscription:
      type: object
      properties:
        userId: { type: string }
        planId: { type: string }
        planName: { type: string }
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CANCELLED, PAST_DUE, TRIALING, NONE]
          description: Status AFTER applying the expiry date; may differ from storedStatus.
        storedStatus: { type: string, nullable: true }
        startDate: { type: string, nullable: true }
        expiryDate: { type: string, nullable: true }
        expiredSinceLastWrite:
          type: boolean
          description: True when a stored ACTIVE was downgraded here because the date passed.
        daysRemaining: { type: integer, nullable: true }
        entitlements:
          type: object
          additionalProperties: { type: boolean }
    ResolvedEntitlement:
      type: object
      properties:
        key: { type: string }
        released: { type: boolean, description: "FF_* release flag — is it built and shipped?" }
        entitled: { type: boolean, description: "Plan grant — may this tier use it?" }
        enabled: { type: boolean, description: "released AND entitled — the only field to gate rendering on." }
        reason:
          type: string
          enum: [ok, not-released, not-in-plan, both]
...